finish logic for getting all news (sorted)

+++ StoronnimV/StoronnimV.DTO/Responses/NewsPage/NewsResponse.cs
@@ -1,8 +1,8 @@
 namespace StoronnimV.DTO.Responses.NewsPage;
 
-public class NewsResponse(int id, string photo, string title, string description, string priority, string date)
+public class NewsResponse(long id, string photo, string title, string description, string priority, string date)
 {
-    public int Id { get; set; } = id;
+    public long Id { get; set; } = id;
     public string Photo { get; set; } = photo;
     public string Title { get; set; } = title;
     public string Description { get; set; } = description;
